Ship modern JS to npm
Previously we transpiled down to IE11 and other older browsers, but we
hadn’t set up babel polyfills. Most notably, this meant that the
compiled code referenced a `regeneratorRuntime` variable that wasn’t
defined (it’s supposed to be defined globally through babel polyfills).
In projects using Create React App we hadn’t noticed this, because CRA
runs npm package code through babel, injecting `regeneratorRuntime`.

However, when trying to use django-bananas in Next.js, the undeclared
`regeneratorRuntime`s threw errors. So what to do?

We _could_ include polyfills in django-bananas. But it might be
unexpected that using django-bananas modifies global objects. And both
we and the user of django-bananas might try to assign
`window.regeneratorRuntime` to different versions.

Another way would be to configure babel to use non-global-polluting
polyfills. But that would make the package unnecessarily heavy
considering that many times admins are used in modern browsers only.

This commit changes the babel setup to use the "esmodules" target, which
targets browsers supporting ES Modules. In other words – the evergreen
browsers. This means that `async`/`await` is preserved – no more
`regeneratorRuntime`. It is then up to the user of django-bananas to
transpile further and add polyfills if deeper browser support is wanted.
Create React App is an example of a build pipeline that supports
transforming code from npm packages.

`import`/`export` are still transpiled to `require`/`module.exports` to
support Next.js server side rendering.