Adjust the way routes are graded.
Before this edit routing attempts to:
/example/fruit/create/
would fire a request to:
/example/fruit/{id}

This was happening due to the routes being resolved in the wrong order.
Now we grade the routes by both segment complexity and alphabetical order.
Since "{" is considered less than anything [a-zA-Z] routes containing parameters
will be resolved after alphabetical, "pure" routes.